client: don't advertise keep-alive when sending error replies
Commit 15346de ("client: Always close connection with request body in case
of error") added logic to close keep alive connections on HTTP errors due
to unconsumed request body data.

However, since the check happens after emitting the standard HTTP headers,
uhttpd might incorrectly reply with a `Connection: keep-alive` even if it
is going to close the connection.

Move the check before the emitting of the response headers in order to
ensure that we're sending the correct `Connection: close` line.

Fixes: 15346de ("client: Always close connection with request body in case of error")
Signed-off-by: Jo-Philipp Wich <jo@mein.io>